requirements.txt
pyTelegramBotAPI
requests
flask
google-generativeai
gunicorn
